Delete record modal is now its own file
