+++ client/src/containers/EditRecord/EditRecord.tsx
@@ -3,10 +3,7 @@ import { useParams, useLocation, useHistory } from 'react-router-dom';
 import { TabNav, Button } from '@ctoec/component-library';
 import AuthenticationContext from '../../contexts/AuthenticationContext/AuthenticationContext';
 import { apiGet, apiDelete } from '../../utils/api';
-import {
-  Child,
-  ReportingPeriod,
-} from '../../shared/models';
+import { Child } from '../../shared/models';
 import { BackButton } from '../../components/BackButton';
 import {
   FamilyIncomeForm,
@@ -16,8 +13,8 @@ import {
   EnrollmentFundingForm,
 } from './Forms';
 import { WithdrawForm } from './Forms/Withdraw/Form';
-import Modal from 'react-modal';
 import { useReportingPeriods } from '../../hooks/useReportingPeriods';
+import { DeleteRecord } from './Forms/DeleteRecord';
 
 const TAB_IDS = {
   CHILD: 'child',
@@ -74,22 +71,6 @@ const EditRecord: React.FC = () => {
     }).then((_rowData) => setRowData(_rowData));
   }, [accessToken, childId, refetch]);
 
-  function deleteRecord() {
-    setIsDeleting(true);
-    apiDelete(`children/${childId}`, { accessToken })
-      // TODO: Swap this total hack out for the roster page
-      // once we have that implemented
-      .then(() => history.push('/check-data/1'))
-
-      .catch((err) => {
-        console.log(err);
-      })
-      .finally(() => {
-        setDeleteModalOpen(false);
-        setIsDeleting(false);
-      });
-  }
-
   if (!rowData) {
     return <></>;
   }
@@ -135,49 +116,11 @@ const EditRecord: React.FC = () => {
             text="Delete record"
             className="margin-right-0"
           />
-          <Modal
+          <DeleteRecord
+            child={rowData}
             isOpen={deleteModalOpen}
-            onRequestClose={toggleDeleteModal}
-            shouldCloseOnEsc={true}
-            shouldCloseOnOverlayClick={true}
-            contentLabel="Delete Modal"
-            // Use style to dynamically trim the bottom to fit the
-            // message, then center in middle of form
-            style={{
-              content: { bottom: 'auto', transform: 'translate(0%, 100%)' },
-            }}
-          >
-            <div className="grid-container">
-              <div className="grid-row margin-top-2">
-                <h2>
-                  Do you want to delete the enrollment for {rowData.firstName}{' '}
-                  {rowData.lastName}?
-                </h2>
-              </div>
-              <div className="grid-row margin-top-2">
-                <span>
-                  Deleting an enrollment record will permanently remove all of
-                  its data
-                </span>
-              </div>
-              <div className="margin-top-4">
-                <div className="grid-row flex-first-baseline space-between-4">
-                  <Button
-                    appearance="outline"
-                    onClick={toggleDeleteModal}
-                    text="No, cancel"
-                  />
-                  <Button
-                    appearance={isDeleting ? 'outline' : 'default'}
-                    onClick={deleteRecord}
-                    text={
-                      isDeleting ? 'Deleting record...' : 'Yes, delete record'
-                    }
-                  />
-                </div>
-              </div>
-            </div>
-          </Modal>
+            toggleOpen={toggleDeleteModal}
+          />
         </div>
       </div>
       <TabNav
